Add scaling to avatar cloning shader
Scaled clones need to be manually offset so their feet are located at
the
ground. Add a float and a toggle for this (_Clone_Av_Height_Fix).

Add ability to rotate clones along with the avatar, which is desirable
in some
situations.

Add 1/100 scale dx parameter, useful for fine-tuning position in game.